safari-600.7-branch/LayoutTests/platform/win/accessibility/select-element-role-expected.txt

This tests that select elements have the correct role.



PASS accessibilityController.focusedElement.role is "combo box"
